requirements.txt
# Core dependencies
python>=3.9

# Data processing
pandas==2.1.3
numpy==1.24.3

# Database
psycopg2-binary==2.9.7
sqlalchemy==1.4.54

# Machine Learning
scikit-learn==1.3.0
joblib==1.3.2

# Stream Processing
kafka-python==2.0.2

# Workflow Orchestration
apache-airflow==2.7.0
apache-airflow-providers-postgres==5.6.0

# Dashboard & Visualization
streamlit==1.28.0
plotly==5.17.0

# Data generation
Faker==19.6.2

# Development
jupyter==1.0.0
ipython==8.15.0
black==23.9.0
flake8==6.1.0
pytest==7.4.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1